requirements.txt
alembic==1.0.11
apache-airflow==1.10.4
apispec==2.0.2
appnope==0.1.0
attrs==19.1.0
awscli==1.16.203
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.193
botocore==1.12.193
cached-property==1.5.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.30
cycler==0.10.0
datadog==0.28.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
docutils==0.14
dumb-init==1.2.2
entrypoints==0.3
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-JWT-Extended==3.21.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
fsspec==0.4.1
funcsigs==1.0.0
future==0.16.0
gunicorn==19.9.0
idna==2.8
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
json-merge-patch==0.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
keyring==19.0.2
kiwisolver==1.1.0
koalas==0.13.0
lazy-object-proxy==1.4.1
lockfile==0.12.2
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.4.1
marshmallow-sqlalchemy==0.17.0
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.16.4
ordereddict==1.1
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pendulum==1.4.4
pexpect==4.7.0
pickleshare==0.7.5
pkginfo==1.5.0.1
prison==0.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
ptyprocess==0.6.0
py4j==0.10.7
pyarrow==0.14.1
pyasn1==0.4.5
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.1
pyrsistent==0.15.3
pyspark==2.4.3
python-daemon==2.1.2
python-dateutil==2.8.0
python-decouple==3.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.1
pytzdata==2019.2
PyYAML==5.1
pyzmq==18.0.2
qtconsole==4.5.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
rsa==3.4.2
s3fs==0.3.3
s3transfer==0.2.1
Send2Trash==1.5.0
setproctitle==1.1.10
six==1.12.0
SQLAlchemy==1.3.6
tabulate==0.8.3
tenacity==4.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
thrift==0.11.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
twine==1.13.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.5.0
WTForms==2.2.1
zope.deprecation==4.4.0
